+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2024-11-13
+
+- Project Level Configuration Support
+- Light theme and Dark theme Toggle
+- Agent icons
+- skeleton loading in agent & server
+
 ## [0.0.2] - 2024-11-12
 - Copilot CLI Support
 - Disable non installed ai agent & ui
@@ -19,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - remove mcp start stop toast
 - reduce caching time to 2 minutes for regular update (from 24 hours)
 - improve ui performance
+- Fix MCP CLI toast on startup
 
 
 ## [0.0.1] - 2024-11-09